pyproject.toml
[project]
name = "task-manager-api"
version = "0.1.0"
description = "Task Manager API - Take-home assignment"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "pydantic-settings>=2.11.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "sqlmodel>=0.0.37",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
